switch from inferRadialDomain to inferRadialRange
inferRadialDomain relies on the heuristic h25 = "median of 25% quantiles of the r channel across marks".
It sets the domain [0, h25] with a default range of [0, 3].

The problem is that "domain" and "range" carry little information about the actual extent of the values (in data or pixel space), and unless you know how the sausage is made you can't really guess how to change either domain or range to meet your visual requirements (though fixing the two solves everything).

With this commit, the default domain becomes the actual extent of values in data space, and the default range are the corresponding values in pixel space, based on the h25 heuristic.

Setting the range to [2, 10] (without setting the domain) will make the smallest dot have a radius of 2 and the largest a radius of 10. This is crystal-clear.

Setting the domain alone (without setting the range) does not change the chart at all, but allows the user to change the extent of a potential future radius legend.

In the case of the http://localhost:8008/?test=diamondsCaratPriceDots test, I had to change add an explicit range: [0,3] to the domain: [0, 100], in order to have the set scale (100 values => a dot of radius 3).

In the general case (no indication given by the user), he heuristic is preserved exactly (only precision issues make the test svg change).