first tentative for a layouts option
two layouts are available:
- "darker" changes the fill channel
- "collide" collides the x/y/r channels

The layouts can be given as a function, which receives the *scaled* values, and returns a function of (index, dimensions), which is called for each facet. You can pass several layouts which will be called one after the other.

(Very rough: no error checking, mutation of the values arrays… no options…)