Trying to fix a "sort fy by x" bug
When we have "sort fy by x" and a transform, the series are not indexed properly.
The facet channel (XV) is indexed by the original data, and the x channel (YV) is indexed by the transformed (grouped) data.

It can be solved, in the case of this unit test, by grouping data by series:
> const XV = x === "fy" || x === "fx" ? [...new Set(X[1].value)] : X[1].value;

however this is not a proper fix, since it should apply only when data is indeed grouped this way.