fix edge case with dodge
align circle edges

filter instead of sort

Avoid a floating point error when computing the dodge transform with fixed radius. Add a line to the cars-dodge example to show how it's "based"

comment for the funky k=2

small tolerance to avoid floating point errors when a circle doesn't fit *exactly*.

Since we're in pixel space it's guaranteed that 1e-6 is "small enough" (and if this doesn't work for someone it's possible to add 0.01 or something to the padding). This approach gives a more compact result in all the test plots.