Introduces a projection.clip option.
Defaults to "frame", which uses the clipRectangle post-clip stream; this has two benefits:
1. the user doesn’t have to add an explicit clip:"frame" option to the geo marks to avoid bleeding in the margins and between facets
2. the computed paths do not extend (invisibly) as much, which seems to make the SVG lighter, and in particular consume less memory when a map zoomed to a large scale has a "heavy" geo mark than spans the whole globe

The second option is "sphere", which can be used to add a clip-path to all the geo marks of the plot (except, of course, the sphere mark itself.

Use null to disable.

Since this changes almost all the test maps, I'm taking this as an opportunity to switch most of them from 50m to 110m geometries, in order to make the tests a bit faster, and the repo a bit smaller.