fix barycentric extrapolation variant (#1705)
* An exact algorithm, which doesn't make use of a delaunay search

* repeat const

* fix a floating point precision issue

When points on the hull are collinear, the extrapolation fails in some regions. I've found a way to fix this by reprojecting the points (to make the hull slightly bulge out, which resolves the ties), but in the end it was too much code for a use case that is pretty bad anyway (the delaunay triangulation itself being unstable in that case). This seems to be quite enough.

* test barycentric data binding

---------

Co-authored-by: Mike Bostock <mbostock@gmail.com>